+++ test/DotRecast.Detour.Test/FindNearestPolyTest.cs
@@ -21,10 +21,12 @@ using NUnit.Framework;
 
 namespace DotRecast.Detour.Test;
 
-
 public class FindNearestPolyTest : AbstractDetourTest
 {
-    private static readonly long[] POLY_REFS = { 281474976710696L, 281474976710773L, 281474976710680L, 281474976710753L, 281474976710733L };
+    private static readonly long[] POLY_REFS =
+    {
+        281474976710696L, 281474976710773L, 281474976710680L, 281474976710753L, 281474976710733L
+    };
 
     private static readonly RcVec3f[] POLY_POS =
     {