+++ CHANGELOG.md
@@ -7,13 +7,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased] - yyyy-mm-dd
 
 ### Added
-- Nothing
+- Added `DtCollectPolysQuery` and `FindCollectPolyTest`
 
 ### Fixed
 - Nothing
 
 ### Changed
 - Changed `IDtPolyQuery` interface to make `Process()` more versatile
+- Changed `PolyQueryInvoker` to `DtActionPolyQuery`
 
 ### Removed
 - Nothing